refactor: EstimatedFood는 vo로 이동(#93)
