+++ src/main/java/capstone/safeat/filter/external/dto/FoodEstimateResponse.java
@@ -1,7 +1,7 @@
 package capstone.safeat.filter.external.dto;
 
 import capstone.safeat.category.domain.Category;
-import capstone.safeat.filter.domain.EstimatedFood;
+import capstone.safeat.filter.vo.EstimatedFood;
 import java.util.List;
 
 public record FoodEstimateResponse(
